Upgrade more packages; auto resolve src for symlinks
