+++ superset-frontend/src/explore/App.jsx
@@ -18,17 +18,9 @@
  */
 import React from 'react';
 import { hot } from 'react-hot-loader/root';
-import { createStore, applyMiddleware, compose } from 'redux';
 import { Provider } from 'react-redux';
-import thunk from 'redux-thunk';
-
-import { initFeatureFlags } from 'src/featureFlags';
-import { initEnhancer } from '../reduxUtils';
-import logger from '../middleware/loggerMiddleware';
 import ToastPresenter from '../messageToasts/containers/ToastPresenter';
 import ExploreViewContainer from './components/ExploreViewContainer';
-import getInitialState from './reducers/getInitialState';
-import rootReducer from './reducers/index';
 
 import setupApp from '../setup/setupApp';
 import setupPlugins from '../setup/setupPlugins';
@@ -38,20 +30,7 @@ import '../../stylesheets/reactable-pagination.less';
 setupApp();
 setupPlugins();
 
-const exploreViewContainer = document.getElementById('app');
-const bootstrapData = JSON.parse(
-  exploreViewContainer.getAttribute('data-bootstrap'),
-);
-initFeatureFlags(bootstrapData.common.feature_flags);
-const initState = getInitialState(bootstrapData);
-
-const store = createStore(
-  rootReducer,
-  initState,
-  compose(applyMiddleware(thunk, logger), initEnhancer(false)),
-);
-
-const App = () => (
+const App = ({ store }) => (
   <Provider store={store}>
     <div>
       <ExploreViewContainer />
