+++ superset-frontend/src/setup/setupApp.ts
@@ -49,7 +49,9 @@ function toggleCheckbox(apiUrlPrefix: string, selector: string) {
 
 export default function setupApp() {
   $(document).ready(function() {
-    $(':checkbox[data-checkbox-api-prefix]').change(function() {
+    $(':checkbox[data-checkbox-api-prefix]').change(function(
+      this: HTMLElement,
+    ) {
       const $this = $(this);
       const prefix = $this.data('checkbox-api-prefix');
       const id = $this.attr('id');
